requirements.txt
flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
mistralai==0.4.0
python-dotenv==1.0.1